Nest EventsCollector in EventsCollectorRule
EventsCollector exists as a separate class even though it is only used
by EventsCollectorRule.

EventsCollector is a test utility class which should only be used in
context of EventsCollectorRule to ensure that EventsCollector objects
are properly torn down after they are initialised.

Let’s refactor EventsCollector by nesting it in EventsCollectorRule.
